[PROJ-1821] Isolate the session response contract
